add roles and permision in controllers

+++ app/Http/Controllers/V1/UserController.php
@@ -3,27 +3,72 @@
 namespace App\Http\Controllers\V1;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\V1\Users\DestroyUserRequest;
+use App\Http\Requests\V1\Files\DestroysFileRequest;
+use App\Http\Requests\V1\Files\IndexFileRequest;
+use App\Http\Requests\V1\Files\UpdateFileRequest;
+use App\Http\Requests\V1\Files\UploadFileRequest;
+use App\Http\Requests\V1\Images\IndexImageRequest;
+use App\Http\Requests\V1\Images\UploadImageRequest;
+use App\Http\Requests\V1\Users\DestroysUserRequest;
+use App\Http\Requests\V1\Users\IndexUserRequest;
 use App\Http\Requests\V1\Users\StoreUserRequest;
 use App\Http\Requests\V1\Users\UpdateUserRequest;
 use App\Http\Resources\V1\Users\UserCollection;
 use App\Http\Resources\V1\Users\UserResource;
+use App\Models\Catalogue;
+use App\Models\File;
 use App\Models\User;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('role:admin');
+        $this->middleware('permission:view-users')->only(['index', 'show']);
+        $this->middleware('permission:store-users')->only(['store']);
+        $this->middleware('permission:update-users')->only(['update']);
+        $this->middleware('permission:delete-users')->only(['destroy', 'destroys']);
+
+        // $this->middleware('role:client');
+        // $this->middleware('permission:view-shops')->only(['index', 'show']);
+        // $this->middleware('permission:store-users')->only(['store']);
+        // $this->middleware('permission:update-users')->only(['update']);
+        // $this->middleware('permission:delete-travels')->only(['destroy', 'destroys']);
+
+        // $this->middleware('role:driver');
+        // $this->middleware('permission:view-travels')->only(['index', 'show']);
+        // $this->middleware('permission:store-users')->only(['store']);
+        // $this->middleware('permission:update-users')->only(['update']);
+        // $this->middleware('permission:delete-users')->only(['destroy', 'destroys']);
+
+        // $this->middleware('role:seller');
+        // $this->middleware('permission:view-shops')->only(['index', 'show']);
+        // $this->middleware('permission:store-products')->only(['store']);
+        // $this->middleware('permission:update-products')->only(['update']);
+        // $this->middleware('permission:update-users')->only(['update']);
+        // $this->middleware('permission:delete-products')->only(['destroy', 'destroys']);
+    }
+
     /**
      * Display a listing of the resource.
-     * @return \Illuminate\Http\Response
      *
      * @return UserCollection
      */
-    public function index()
+    public function index(IndexUserRequest $request)
     {
-        return (new UserCollection(User::paginate()))
+        $sorts = explode(',', $request->sort);
+
+        $users = User::customSelect($request->fields)->customOrderBy($sorts)
+            ->name($request->input('name'))
+            ->lastname($request->input('lastname'))
+            ->paginate();
+
+        return (new UserCollection($users))
             ->additional([
                 'msg' => [
-                    'summary' => 'consulta exitosa',
+                    'summary' => 'success',
                     'detail' => '',
                     'code' => '200'
                 ]
@@ -34,23 +79,35 @@ class UserController extends Controller
      * Store a newly created resource in storage.
      *
      * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\JsonResponse
+     * @return UserResource
      */
     public function store(StoreUserRequest $request)
     {
         $user = new User();
-        $user->name = $request->input('name');
-        $user->phone = $request->input('phone');;
-        $user->email = $request->input('email');
+        $user->identificationType()->associate(Catalogue::find($request->input('identificationType.id')));
+        $user->sex()->associate(Catalogue::find($request->input('sex.id')));
+        $user->gender()->associate(Catalogue::find($request->input('gender.id')));
+        $user->bloodType()->associate(Catalogue::find($request->input('bloodType.id')));
+        $user->ethnicOrigin()->associate(Catalogue::find($request->input('ethnicOrigin.id')));
+        $user->civilStatus()->associate(Catalogue::find($request->input('civilStatus.id')));
+
+        $user->username = $request->input('username');
         $user->password = $request->input('password');
-        $user->direction = $request->input('direction');
-        $user->genred = $request->input('genred');
-        $user->save();
+        $user->name = $request->input('name');
+        $user->lastname = $request->input('lastname');
+        $user->birthdate = $request->input('birthdate');
+        $user->email = $request->input('email');
+
+        DB::transaction(function () use ($request, $user) {
+            $user->save();
+            $user->addPhones($request->input('phones'));
+            $user->addEmails($request->input('emails'));
+        });
 
         return (new UserResource($user))
             ->additional([
                 'msg' => [
-                    'summary' => 'creacion correcta',
+                    'summary' => 'Usuario Creado',
                     'detail' => '',
                     'code' => '200'
                 ]
@@ -65,12 +122,11 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        //metodo find()
         return (new UserResource($user))
             ->additional([
                 'msg' => [
                     'summary' => 'success',
-                    'detail' => 'consulta exitosa',
+                    'detail' => '',
                     'code' => '200'
                 ]
             ]);
@@ -81,21 +137,31 @@ class UserController extends Controller
      *
      * @param \Illuminate\Http\Request $request
      * @param \App\Models\User $user
-     * @return \Illuminate\Http\JsonResponse
+     * @return UserResource
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $user->identificationType()->associate(Catalogue::find($request->input('identificationType.id')));
+        $user->sex()->associate(Catalogue::find($request->input('sex.id')));
+        $user->gender()->associate(Catalogue::find($request->input('gender.id')));
+        $user->bloodType()->associate(Catalogue::find($request->input('bloodType.id')));
+        $user->ethnicOrigin()->associate(Catalogue::find($request->input('ethnicOrigin.id')));
+        $user->civilStatus()->associate(Catalogue::find($request->input('civilStatus.id')));
+
+        $user->username = $request->input('username');
         $user->name = $request->input('name');
-        $user->phone = $request->input('phone');;
+        $user->lastname = $request->input('lastname');
+        $user->birthdate = $request->input('birthdate');
         $user->email = $request->input('email');
-        $user->password = $request->input('password');
-        $user->direction = $request->input('direction');
-        $user->genred = $request->input('genred');
+
         $user->save();
+        $user->addPhones($request->input('phones'));
+        $user->addEmails($request->input('emails'));
+
         return (new UserResource($user))
             ->additional([
                 'msg' => [
-                    'summary' => 'actualizacion correcta',
+                    'summary' => 'Usuario Actualizado',
                     'detail' => '',
                     'code' => '200'
                 ]
@@ -106,34 +172,99 @@ class UserController extends Controller
      * Remove the specified resource from storage.
      *
      * @param \App\Models\User $user
-     * @return \Illuminate\Http\JsonResponse
+     * @return UserResource|\Illuminate\Http\JsonResponse
      */
-    public function destroy($user)
+    public function destroy(Request $request, User $user)
     {
-        $user = User::find($user);
+        if ($request->user()->id === $user->id) {
+            return response()->json([
+                'msg' => [
+                    'summary' => 'No se puede eliminar',
+                    'detail' => 'No puede eliminar el usuario logueado',
+                    'code' => '400'
+                ],
+            ], 400);
+        }
         $user->delete();
-
         return (new UserResource($user))
             ->additional([
                 'msg' => [
-                    'summary' => 'usuario eliminado correcta',
+                    'summary' => 'Usuario Eliminado',
                     'detail' => '',
                     'code' => '200'
                 ]
             ]);
     }
 
-    public function destroys(DestroyUserRequest $request)
+    public function destroys(DestroysUserRequest $request)
     {
+        if (in_array($request->user()->id, $request->ids)) {
+            return response()->json([
+                'msg' => [
+                    'summary' => 'No se pudeo eliminar',
+                    'detail' => 'El usuario se encuentra logueado',
+                    'code' => '400'
+                ],
+            ], 400);
+        }
+
+        $users = User::whereIn('id', $request->input('ids'))->get();
         User::destroy($request->input('ids'));
 
-        return (new UserResource($request))
+        return (new UserCollection($users))
             ->additional([
                 'msg' => [
-                    'summary' => ' eliminacion exitosa',
+                    'summary' => 'Usuarios Eliminados',
                     'detail' => '',
-                    'code' => '200'
+                    'code' => '201'
                 ]
             ]);
+    }
+
+    // Images
+    public function uploadImage(UploadImageRequest $request, User $user)
+    {
+        return $user->uploadImage($request);
+    }
+
+    public function indexImage(IndexImageRequest $request)
+    {
+        return (new ImageController())->index($request, User::find($request->input('id')));
+    }
+
+    // Files
+    public function indexFiles(IndexFileRequest $request, User $user)
+    {
+        return $user->indexFiles($request);
+    }
+
+    public function uploadFile(UploadFileRequest $request, User $user)
+    {
+        return $user->uploadFile($request);
+    }
+
+    public function downloadFile(User $user, File $file)
+    {
+        return $user->downloadFile($file);
+    }
+
+    public function showFile(User $user, File $file)
+    {
+        return $user->showFile($file);
+    }
+
+    public function updateFile(UpdateFileRequest $request, User $user, File $file)
+    {
+        return $user->updateFile($request, $file);
+    }
+
+    public function destroyFile(User $user, File $file)
+    {
+        return $user->destroyFile($file);
+    }
+
+    public function destroyFiles(User $user, DestroysFileRequest $request)
+    {
+        return $user->destroyFiles($request);
     }
 }
